+++ shaders/map_height.hlsl
@@ -9,12 +9,12 @@ float ComputeVertexHeight( float4 p_vpos, float4 p_uv, float p_plains_amplitude,
 
 	float res = 0.0;
 
-	double n_vpos_x = ( p_vpos.x / 2.0 ) + 0.5;
-	double n_vpos_y = ( p_vpos.y / 2.0 ) + 0.5;
-	double n_vpos_z = ( p_vpos.z / 2.0 ) + 0.5;
+	float n_vpos_x = ( p_vpos.x / 2.0 ) + 0.5;
+	float n_vpos_y = ( p_vpos.y / 2.0 ) + 0.5;
+	float n_vpos_z = ( p_vpos.z / 2.0 ) + 0.5;
 
 
-	double3 f2;
+	float3 f2;
 	f2[0] = lerp( -fbm_mountains_hl, fbm_mountains_hl, n_vpos_x );
 	f2[1] = lerp( -fbm_mountains_hl, fbm_mountains_hl, n_vpos_y );
 	f2[2] = lerp( -fbm_mountains_hl, fbm_mountains_hl, n_vpos_z );
@@ -26,7 +26,7 @@ float ComputeVertexHeight( float4 p_vpos, float4 p_uv, float p_plains_amplitude,
 	}
 
 
-	double3 f3;
+	float3 f3;
 	f3[0] = lerp( -fbm_uvnoise_hl, fbm_uvnoise_hl, n_vpos_x );
 	f3[1] = lerp( -fbm_uvnoise_hl, fbm_uvnoise_hl, n_vpos_y );
 	f3[2] = lerp( -fbm_uvnoise_hl, fbm_uvnoise_hl, n_vpos_z );
@@ -53,7 +53,7 @@ float ComputeVertexHeight( float4 p_vpos, float4 p_uv, float p_plains_amplitude,
 	float pn2 = 9.0 * pn;
 	float weight = exp( pn2 ) / 1000.0;
 	
-	double hl = color.y;
+	float hl = color.y;
 	res = lerp( plain, plain + mountains, hl * weight );
 	
     res += p_offset;
